reworked business logic

+++ src/main/java/data/entities/UserBO.java
@@ -45,8 +45,8 @@ public class UserBO {
 
     @Temporal(TemporalType.DATE)
     private Date birthday;
-    private double height;
-    private double weight;
+    private int height;
+    private int weight;
 
     @OneToMany
     private List<ChallengeBO> challenges;
@@ -197,19 +197,19 @@ public class UserBO {
         this.birthday = birthday;
     }
 
-    public double getHeight() {
+    public int getHeight() {
         return height;
     }
 
-    public void setHeight(double height) {
+    public void setHeight(int height) {
         this.height = height;
     }
 
-    public double getWeight() {
+    public int getWeight() {
         return weight;
     }
 
-    public void setWeight(double weight) {
+    public void setWeight(int weight) {
         this.weight = weight;
     }
